+++ frontend/src/assets/animations/fox_hello.json
