+++ frontend/src/assets/animations/fox_programmer.json
